Revert the change to DNS_TTL and fix the renewer's TTL
In a previous change, DNSConstants.DNS_TTL had been changed to reduce
the TTL of mDNS services advertised by the A@H broker.  Instead of
changing this for all users of JmDNS, A@H will use the
DNSStateTask.setDefaultTTL internal API to control the TTL.

Also, patch the Renewer so that the renewal interval is based on the
task's own TTL instead of the hardcoded TTL in DNSConstants.

Change-Id: I9f42b98e8088267b38d380edfd8b2e737d00414e